+++ NEWS.md
@@ -1,3 +1,15 @@
+# bigQueryR 0.5.0
+
+* Support listing more than 50 datasets in `bqr_list_dataset`
+* Change `bqr_list_tables` to list all tables in a dataset by default
+* Add `bqr_copy_dataset`
+* Add `Table` and `bqr_update_table`
+* Support uploading nested lists via `toJSON`
+* Add writeDisposition to table loads
+* Allow creation of empty tables
+* Supporting supplying SQL via a file ending with `.sql` for `bqr_query()`
+* Update to new `googleAuthR>1.1.1`
+
 # bigQueryR 0.4.0
 
 * support `nullMarker`, `maxBadRecords`, `fieldDelimiter` in upload jobs